HOCS-4821: check increment value
Early failure check on the increment value against what the accepted SemVer identifiers are. Fails the build if it's not an acceptable value.

Change to the SemVer tag generation to first check the base reference for the pull request against existing tags. This tag is then incremented by the identifier and checked against the existing tags to ensure uniqueness. If the tag already exists the build is failed.

Add a helper method to retrieve all of the action inputs at once and return them as a single object.